Rename `bazel_features.bzlmod` to `bazel_features.repos`
Continuing the trend of de-emphasizing the name "Bzlmod". I think it would make more sense to group features related to external dependencies together, instead of just Bzlmod features. Not sure whether this should be called `external` or `external_deps` or just `repos`, so I just went with the shortest one.