Add has_parallel_download feature to detect if rctx.download supports parallel downloads (#29)
Coming in 7.1.0,
[repository_ctx#download](https://bazel.build/rules/lib/builtins/repository_ctx#download)
has a new argument, block:

> If set to false, the call returns immediately and instead of the
regular return value, it returns a token with one single method, wait(),
which blocks until the download is finished and returns the usual return
value or throws as usual.

It would be useful to be able to update repository rules to use this
feature when available.